+++ src/promptflow/tests/sdk_cli_azure_test/conftest.py
@@ -16,6 +16,7 @@ from azure.ai.ml import MLClient
 from azure.ai.ml.constants._common import AZUREML_RESOURCE_PROVIDER, RESOURCE_ID_FORMAT
 from azure.ai.ml.entities import Data
 from azure.core.exceptions import ResourceNotFoundError
+from dotenv import load_dotenv
 from pytest_mock import MockerFixture
 
 from promptflow._telemetry.telemetry import TELEMETRY_ENABLED
@@ -23,20 +24,17 @@ from promptflow._utils.utils import environment_variable_overwrite
 from promptflow.azure import PFClient
 
 from ._azure_utils import get_cred
-from .recording_utilities import (
-    PFAzureIntegrationTestRecording,
-    get_pf_client_for_playback,
-    is_live,
-    is_live_and_not_recording,
-)
+from .recording_utilities import PFAzureIntegrationTestRecording, get_pf_client_for_replay, is_live, is_replay
 
 FLOWS_DIR = "./tests/test_configs/flows"
 DATAS_DIR = "./tests/test_configs/datas"
 
+load_dotenv()
+
 
 @pytest.fixture
 def tenant_id() -> str:
-    if not is_live():
+    if is_replay():
         return ""
     credential = get_cred()
     access_token = credential.get_token("https://management.azure.com/.default")
@@ -63,8 +61,8 @@ def ml_client(
 
 @pytest.fixture
 def remote_client() -> PFClient:
-    if not is_live():
-        yield get_pf_client_for_playback()
+    if is_replay():
+        yield get_pf_client_for_replay()
     else:
         # enable telemetry for CI
         with environment_variable_overwrite(TELEMETRY_ENABLED, "true"):
@@ -85,8 +83,8 @@ def remote_workspace_resource_id() -> str:
 
 @pytest.fixture
 def remote_client_int() -> PFClient:
-    if not is_live():
-        yield get_pf_client_for_playback()
+    if is_replay():
+        yield get_pf_client_for_replay()
     else:
         # enable telemetry for non-playback CI
         with environment_variable_overwrite(TELEMETRY_ENABLED, "true"):
@@ -199,12 +197,17 @@ def flow_serving_client_remote_connection(mocker: MockerFixture, remote_workspac
 
 @pytest.fixture(scope="function")
 def vcr_recording(request: pytest.FixtureRequest, tenant_id: str) -> PFAzureIntegrationTestRecording:
+    """Fixture to record or replay network traffic.
+
+    If the test mode is "record" or "replay", this fixture will locate a YAML (recording) file
+    based on the test file, class and function name, write to (record) or read from (replay) the file.
+    """
     recording = PFAzureIntegrationTestRecording.from_test_case(
         test_class=request.cls,
         test_func_name=request.node.name,
         tenant_id=tenant_id,
     )
-    if not is_live_and_not_recording():
+    if not is_live():
         recording.enter_vcr()
         request.addfinalizer(recording.exit_vcr)
     yield recording
@@ -224,7 +227,7 @@ def randstr(vcr_recording: PFAzureIntegrationTestRecording) -> Callable[[str], s
 # we expect this fixture only work when running live test without recording
 # when recording, we don't want to record any application insights secrets
 # when replaying, we also don't need this
-@pytest.fixture(autouse=not is_live_and_not_recording())
+@pytest.fixture(autouse=not is_live())
 def mock_appinsights_log_handler(mocker: MockerFixture) -> None:
     dummy_logger = logging.getLogger("dummy")
     mocker.patch("promptflow._telemetry.telemetry.get_telemetry_logger", return_value=dummy_logger)
@@ -233,6 +236,12 @@ def mock_appinsights_log_handler(mocker: MockerFixture) -> None:
 
 @pytest.fixture
 def single_worker_thread_pool() -> None:
+    """Mock to use one thread for thread pool executor.
+
+    VCR.py cannot record network traffic in other threads, and we have multi-thread operations
+    during resolving the flow. Mock it using one thread to make VCR.py work.
+    """
+
     def single_worker_thread_pool_executor(*args, **kwargs):
         return ThreadPoolExecutor(max_workers=1)
 
@@ -245,12 +254,18 @@ def single_worker_thread_pool() -> None:
 
 @pytest.fixture
 def mock_set_headers_with_user_aml_token(mocker: MockerFixture) -> None:
+    """Mock set aml-user-token operation.
+
+    There will be requests fetching cloud metadata during retrieving AML token, which will break during replay.
+    As the logic comes from azure-ai-ml, changes in Prompt Flow can hardly affect it, mock it here.
+    """
     mocker.patch("promptflow.azure._restclient.flow_service_caller.FlowServiceCaller._set_headers_with_user_aml_token")
     return
 
 
 @pytest.fixture
 def mock_get_azure_pf_client(mocker: MockerFixture, remote_client: PFClient) -> None:
+    """Mock PF Azure client to avoid network traffic during replay test."""
     mocker.patch(
         "promptflow._cli._pf_azure._run._get_azure_pf_client",
         return_value=remote_client,